Try and Fix XDP redirect in KIND
Try and ensure XDP redirect will work in KIND

This Updates the XDP program to rewrite:
- Src Mac address
- Dst Port address

and allows us to configure if we want to run cksum
calculation on a per backend basis. It also fixes
the existing CKSUM calculation function to ignore
trailing bits.

Lastly We should be using the autogenerated bpf2go
map structs from `bpf_*.go`

Signed-off-by: Andrew Stoycos <astoycos@redhat.com>