feat: add grpc api for dataplane
This also includes a new xtask for manually
inserting and removing routing data from the
dataplane, e.g.:

  cargo xtask grpc-client --vip-ip 172.17.0.2 --port 9875

NOTE: This change was made in lieu of adding a
pull mechanism for dataplane configuration (which
is actually what I added before in a previous
iteration). The reason for this choice was purely
to make development easy in the coming weeks,
allowing a quick and easy way to insert routing
data into the dataplane for debugging and testing.
I would not be surprised if later once things are
futher along we went the other way again.