feat: performance test scaffolding (#72)
This partially resolves #44 by adding a performance testing suite and
some tooling for performance testing based on `iperf3`. An image build
for `iperf3` is provided and hosted here as the upstream project is not
interested in providing one themselves. This also includes some
refactoring and moving of test code into common areas.

We're blocked on going any further until #15 is completed as `iperf3`
requires TCP for metadata exchanges so I figured we could just get this
stuff in for the moment so that we have less huge PRs.