controlplane: improve garbage collection upon deletion
Avoid fetching Endpoints while deleting the Targets in the dataplane,
since we only need the Gateway's IP addr and port for such an action.
This avoids potential errors that can arise due to an Endpoint being
deleted before the reconciler reacts to a UDPRoute/TCPRoute's deletion,
which can block the object from being garbage collected forever due to
the presence of finalizers.

Signed-off-by: Sanskar Jaiswal <jaiswalsanskar078@gmail.com>