tcproute: add support for round-robin load balancing
Add support for round-robin load balancing for `TCPRoute`. This enables
mutliple backend references in a UDPRoute object, with traffic being
distributed to each backend in a round-robin fashion.

A new BPF map `TCP_CONNECTIONS` was introduced to help keep track of
active TCP connections, with its key storing the client <ip:port>
identifier and the value storing the backend's <ip:port> along with
the state of the connection.

Signed-off-by: Sanskar Jaiswal <jaiswalsanskar078@gmail.com>