Ensure rust code passes code standards
Adds a CI suite for building and linting rust code in the
dataplane. This minimizes and/or altogher blocks
non-idiomatic rust code from sneaking into the code base.

NOTE: Builds happen on a directory by directory basis and
are explicitly written as steps. Once we migrate the controller
from Go to Rust, we'll need to add a step for building that
as well.

Signed-off-by: Evan Jones <evan.a.jones3@gmail.com>